+++ bacula/src/jcr.h
@@ -270,7 +270,7 @@ public:
    alist *pki_recipients;             /* Trusted Recipients */
    CRYPTO_SESSION *pki_session;       /* PKE Public Keys + Symmetric Session Keys */
    uint8_t *pki_session_encoded;      /* Cached DER-encoded copy of pki_session */
-   size_t pki_session_encoded_size;   /* Size of DER-encoded pki_session */
+   int32_t pki_session_encoded_size;  /* Size of DER-encoded pki_session */
    POOLMEM *crypto_buf;               /* Encryption/Decryption buffer */
    DIRRES* director;                  /* Director resource */
 #endif /* FILE_DAEMON */
